fix the conflicts
