+++ server/statistics/region_collection_test.go
